+++ arpack-ng/cppbuild.sh
@@ -7,7 +7,7 @@ if [[ -z "$PLATFORM" ]]; then
     exit
 fi
 
-ARPACK_NG_VERSION=3.6.3
+ARPACK_NG_VERSION=3.7.0
 download https://github.com/opencollab/arpack-ng/archive/$ARPACK_NG_VERSION.tar.gz arpack-ng-$ARPACK_NG_VERSION.tar.gz
 
 mkdir -p $PLATFORM
@@ -22,12 +22,12 @@ if [[ "${ACLOCAL_PATH:-}" == C:\\msys64\\* ]]; then
 fi
 patch -Np1 < ../../../arpack-ng-configure.patch || true # bash bootstrap
 chmod 755 configure build-aux/install-sh
-sedinplace 's/std::real(sigma) + std::imag(sigma) \* I/*reinterpret_cast<_Complex double*>(\&sigma)/g' arpack.hpp
-sedinplace 's/std::real(sigma) + _Complex_I \* std::imag(sigma)/*reinterpret_cast<_Complex double*>(\&sigma)/g' arpack.hpp
-sedinplace 's/internal::s/s/g' arpack.hpp
-sedinplace 's/internal::d/d/g' arpack.hpp
-sedinplace 's/internal::cn/cn/g' arpack.hpp
-sedinplace 's/internal::z/z/g' arpack.hpp
+sedinplace 's/std::real(sigma) + std::imag(sigma) \* I/*reinterpret_cast<_Complex double*>(\&sigma)/g' ICB/arpack.hpp
+sedinplace 's/std::real(sigma) + _Complex_I \* std::imag(sigma)/*reinterpret_cast<_Complex double*>(\&sigma)/g' ICB/arpack.hpp
+sedinplace 's/internal::s/s/g' ICB/arpack.hpp
+sedinplace 's/internal::d/d/g' ICB/arpack.hpp
+sedinplace 's/internal::cn/cn/g' ICB/arpack.hpp
+sedinplace 's/internal::z/z/g' ICB/arpack.hpp
 
 OPENBLAS_PATH="$INSTALL_PATH/../../../openblas/cppbuild/$PLATFORM/"