+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+ * Add `GetComponentImagesExample` for Tesseract ([pull #673](https://github.com/bytedeco/javacpp-presets/pull/673))
+ * Add presets for nGraph 0.11.1 ([pull #642](https://github.com/bytedeco/javacpp-presets/pull/642))
+ * Upgrade presets for ARPACK-NG 3.7.0
+ * Build OpenCV without UI when environment variable `HEADLESS=yes` ([pull #667](https://github.com/bytedeco/javacpp-presets/pull/667))
+
+### January 11, 2019 version 1.4.4
  * Bundle the full version of MKL now that its new license permits it ([issue #601](https://github.com/bytedeco/javacpp-presets/issues/601))
  * Bundle libraries from raspberrypi/userland to avoid loading issues on `linux-armhf` devices other than Raspberry Pi ([issue bytedeco/javacv#1118](https://github.com/bytedeco/javacv/issues/1118))
  * Bundle the new official Java/Scala API of MXNet via the `jnimxnet` library